リファクタリングとビルドテスト
